+++ Python/Final project/Tests/test_reg.py
@@ -3,14 +3,13 @@ import os
 import pytest
 from selenium import webdriver
 from Pages.reg_page import RegPage
-from settings import DEFAULT_WIDTH, DEFAULT_HEIGHT, conf_chrome, conf_firefox, conf_safari, conf_edge, rnd_str, \
-    screenshot_folder_name
+from settings import DEFAULT_WIDTH, DEFAULT_HEIGHT, conf_chrome, conf_firefox, conf_safari, conf_edge, rnd_str,\
+    DELETE_UNNEEDED_SCREENSHOTS, ValidData as Valid
 from time import sleep
 
 
 # ADD 'safari' when testing not on Mac
-# @pytest.fixture(params=['chrome', 'firefox', 'edge'], scope='session')  # , 'safari'], scope='session')
-@pytest.fixture(params=['chrome'], scope='session')
+@pytest.fixture(params=['chrome', 'firefox', 'edge'], scope='session')  # , 'safari'], scope='session')
 def init_driver(request, width=DEFAULT_WIDTH, height=DEFAULT_HEIGHT):
     """
     initialize webdriver
@@ -41,14 +40,32 @@ def init_driver(request, width=DEFAULT_WIDTH, height=DEFAULT_HEIGHT):
                          [
                              ('Марья', 'Петровна', 'Владивосток', '+79991235689', 'dxoOd386', 'same', True),
                              ('Максим', 'Квакин', 'rnd', 'max_kvakin@ya.ya.ru', 'boWu2Mx601', 'same', True),
-                             ('Василий', 'Иваныч', 0, '+79112223344', rnd_str(), 'same', False),
-                             ('User', 'Lastname', 'last', 'user@mail.com', rnd_str(), 'same', False),
+                             ('Василий', 'Иваныч', 0, "+79997778899", rnd_str(), 'same', False),
+                             ('User', 'Lastname', 'last', 'boris_berezov@sky.ru', rnd_str(), 'same', False),
                              ('Супер', 'Админ', 'rnd', 'admin@rt.ru', 'passkey', 'same', False),
+                             # ('Семён', '', 'Владивосток', '+79991235689', 'dxoOd386', 'same', False),
+                             # ('Марья', 'Петровна', 'Владивосток', '1@.ru', 'dxoOd386', 'same', False),
+                             # ('Марья', 'Петровна', 'Владивосток', '', 'dxoOd386', 'same', False),
+                             # (rnd_str(256, 'r'), rnd_str(256, 'r'), 'Владивосток', '+79991235689', 'dxoOd386', 'same',
+                             #  False),
+                             # ('Марья', 'Петровна', 'Владивосток', f'+79{rnd_str(20, "d")}', 'dxoOd386', 'same', False),
+                             # ('Марья', 'Петровна', 'Владивосток', rnd_str(256), 'dxoOd386', 'same', False),
+                             # (rnd_str(15, 's'), rnd_str(15, 's'), 'Владивосток', rnd_str(15, 's'), 'dxoOd386', 'same',
+                             #  False),
                          ],
                          ids = [
-                             'Positive registration by phone number', 'Positive registration by email',
-                             'Registration with taken username', 'Negative test with invalid names',
-                             'Negative test with weak password'
+                             'Positive registration by phone number',
+                             'Positive registration by email',
+                             'Negative test with already registered phone number',
+                             'Negative test with invalid names',
+                             'Negative test with weak password',
+                             # 'Negative test with empty first_name',
+                             # 'Negative test with invalid email address',
+                             # 'Negative test with empty username',
+                             # 'Negative test with long names',
+                             # 'Negative test with long phone number',
+                             # 'Negative test with long username',
+                             # 'Negative test with special symbols in user names',
                          ])
 def test_reg_page(init_driver, first_name: str, last_name: str, region, username, password, password2, expectation: bool):
     """Test of RegPage"""
@@ -80,9 +97,10 @@ def test_reg_page(init_driver, first_name: str, last_name: str, region, username
 
     chk = page.check_code_input_container()
     if chk == expectation:
-        for s in screenshots:
-            if os.path.exists(s):
-                os.remove(s)
+        if DELETE_UNNEEDED_SCREENSHOTS:
+            for s in screenshots:
+                if os.path.exists(s):
+                    os.remove(s)
         page.log.append('    ... ok')
     else:
         page.log.append('   ... failed')
